backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
python-multipart==0.0.9
pydantic==2.7.1
pdfplumber==0.11.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.3.0
pytesseract==0.3.10
spacy==3.7.4
scikit-learn==1.4.2
numpy==1.26.4
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
reportlab==4.2.0
anthropic==0.25.1
sentence-transformers==2.7.0
